requirements.txt
# gradio
# sentence-transformers
# faiss-cpu
# pypdf
# transformers
# torch==2.9.1+cpu
# --extra-index-url https://download.pytorch.org/whl/cpu


flask
sentence-transformers
faiss-cpu
pypdf
transformers
torch==2.5.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
